Add .vercel to .gitignore and update Storybook link in README
The .vercel directory has been added to the .gitignore file to prevent Vercel build files from being included in the versioning system. This is important to ensure that only source files are managed via Git and not build artifacts.

Additionally, an vercel URL link has been included in the README file, which is lead to the Storybook documentation.