Update rollup configuration and add release scripts
This commit includes modifications in the rollup configuration file and package.json. The babel plugin was removed from rollup.config.js to resolve conflict with postCSS. Besides, a new configuration has been introduced for 'src/index.ts' with the output format updated to 'es'.

In package.json, new scripts for versioning are introduced to handle major, minor and patch releases efficiently. This aims to streamline the release process and enhance the code maintenance.